umbrella_backend/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.2.12
djangorestframework==3.17.1
idna==3.11
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
requests==2.33.1
resend==2.27.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
